Merges RenderTextureComponent into render_texture This did not need to be a separate file
